Convert the temp table approach in metric_view() to using functions instead.
The function-based approach is very close in performance to the temp table
approach and has the added benefit of not requiring the metric_view() function
to be volatile. Most of these functions mirror existing functions in the
_timescaledb_internal schema. The functions in the _timescaledb_internal schema
are scoped to a single hypertable. These functions return results for all
the hypertables in a given schema at once. The hypertable-at-a-time approach
requires a left lateral join when you need to get results for many tables.
This looping is much slower than running the same calc over a bunch of tables
in a single pass.

New functions:

* hypertable_local_size(schema_name_in name)
* hypertable_node_up(schema_name_in name)
* hypertable_remote_size(schema_name_in name)
* hypertable_compression_stats(schema_name_in name)